pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.7",
    "langchain-chroma>=1.1.0",
    "langchain-core>=1.2.28",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.23",
    "torch>=2.11.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
